Moving MindViewer test to right place
