--- src/main/java/br/unicamp/cst/core/entities/TestCodelet.java
@@ -1,35 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2012  DCA-FEEC-UNICAMP
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the GNU Lesser Public License v3
- * which accompanies this distribution, and is available at
- * http://www.gnu.org/licenses/lgpl.html
- * 
- * Contributors:
- *     K. Raizer, A. L. O. Paraense, R. R. Gudwin - initial API and implementation
- ******************************************************************************/
-
-package br.unicamp.cst.core.entities;
-
-/**
- *
- * @author gudwin
- */
-public class TestCodelet extends Codelet {
-    
-    public TestCodelet(String nome) {
-        setName(nome);
-    }
-    
-        public void accessMemoryObjects() {
-            
-        }
-
-	public void calculateActivation() {
-            
-        }
-
-	public void proc() {
-            
-        }
-    }
