+++ src/test/java/br/unicamp/cst/util/MindViewerTest.java
@@ -1,0 +1,87 @@
+/**
+ * 
+ */
+package br.unicamp.cst.util;
+
+import java.util.ArrayList;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+
+import br.unicamp.cst.core.entities.Codelet;
+import br.unicamp.cst.core.entities.MemoryContainer;
+import br.unicamp.cst.core.entities.MemoryObject;
+import br.unicamp.cst.core.entities.Mind;
+
+/**
+ * @author andre
+ *
+ */
+public class MindViewerTest {
+	
+	@BeforeClass
+    public static void beforeAllTestMethods() {
+    }
+
+	@AfterClass
+    public static void afterAllTestMethods() {
+    }
+    
+    @Test
+    public void testRosTopics() {
+    	
+    	/* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(MindViewer.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(MindViewer.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(MindViewer.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(MindViewer.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+        Mind m = new Mind();
+        MemoryObject m1 = m.createMemoryObject("M1", 0.12);
+        MemoryObject m2 = m.createMemoryObject("M2", 0.32);
+        MemoryObject m3 = m.createMemoryObject("M3", 0.44);
+        MemoryObject m4 = m.createMemoryObject("M4", 0.52);
+        MemoryObject m5 = m.createMemoryObject("M5", 0.12);
+        MemoryContainer m6 = m.createMemoryContainer("C1");
+        MemoryContainer m7 = m.createMemoryContainer("C2");
+        m7.setI(0.55, 0.23);
+        m6.setI(0.33, 0.22);
+        m6.setI(0.12, 0.13);
+        m6.setI(m7);
+        Codelet c = new TestCodelet("Codelet 1");
+        c.addInput(m1);
+        c.addInput(m2);
+        c.addOutput(m3);
+        c.addOutput(m4);
+        c.addBroadcast(m5);
+        m.insertCodelet(c);
+        Codelet c2 = new TestCodelet("Codelet 2");
+        c2.addInput(m4);
+        c2.addInput(m5);
+        c2.addOutput(m6);
+        c2.addOutput(m3);
+        c2.addBroadcast(m5);
+        m.insertCodelet(c2);
+        MindViewer ov = new MindViewer(m, "Mind", new ArrayList<>());
+        ov.setVisible(true);
+        System.out.println("Teste:");
+        // ov.updateTree(m);    	
+    }
+}
