+++ src/test/java/br/unicamp/cst/util/TestCodelet.java
@@ -1,0 +1,34 @@
+/**
+ * 
+ */
+package br.unicamp.cst.util;
+
+import br.unicamp.cst.core.entities.Codelet;
+
+/**
+ * 
+ * @author gudwin
+ *
+ */
+public class TestCodelet extends Codelet{
+
+	public TestCodelet(String name) {
+		setName(name);
+	}
+
+	@Override
+	public void accessMemoryObjects() {
+		
+	}
+
+	@Override
+	public void calculateActivation() {
+		
+	}
+
+	@Override
+	public void proc() {
+		
+	}
+
+}
